backend/requirements.txt
flask
flask-cors
supabase==1.0.33
bcrypt
pyjwt
